pyproject.toml
[tool.poetry]
name = "django-toy"
version = "0.1.0"
description = ""
authors = ["Dam <60846068+j2h30728@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
